+++ src/main/java/org/verapdf/cos/filters/COSFilterRC4DecryptionDefault.java
@@ -74,7 +74,7 @@ public class COSFilterRC4DecryptionDefault extends ASBufferedInFilter {
     @Override
     public int read(byte[] buffer, int off, int size) throws IOException {
         if(this.bufferSize() == 0) {
-            int bytesFed = (int) this.feedBuffer(getBufferCapacity());
+            int bytesFed = this.feedBuffer(getBufferCapacity());
             if (bytesFed == -1) {
                 return -1;
             }
